backend/requirements.txt
# Backend Requirements
# Python 3.11+

# Django Framework
django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.27.0

# CORS and Security
django-cors-headers>=4.3.0
python-dotenv>=1.0.0

# Image Processing
Pillow>=10.0.0

# PDF Generation
reportlab>=4.0.0

# HTTP Requests
requests>=2.32.0

# Database (Production)
# PostgreSQL for Azure/Production
psycopg2-binary>=2.9.9

# Production Server & Static Files
gunicorn>=21.2.0
whitenoise>=6.6.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
